requirements.txt
google-adk==1.0.0
streamlit==1.32.0
feedparser==6.0.10
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.11.4
